Update maven.md to correct numbering and indents
Step 2 was missing, so I have shifted 3 and 4 up.

Code blocks on the [site](http://adobe-consulting-services.github.io/acs-aem-commons/pages/maven.html) were incorrectly indented.  It seems the MD parser was trimming the first line of the blocks.  Reduced the indent in each block to compensate for this.  Another option may be to instead use
```xml
<fenced>code</fenced>
```